ais/requirement.txt
paho-mqtt==2.1.0
pyais==2.7.2
